FIX:MAXFILE i32 to usize

+++ kernel-rs/src/fs.rs
@@ -407,7 +407,7 @@ impl Inode {
         if off > (*self).size || off.wrapping_add(n) < off {
             return -1;
         }
-        if off.wrapping_add(n) as usize > (MAXFILE as usize).wrapping_mul(BSIZE) {
+        if off.wrapping_add(n) as usize > MAXFILE.wrapping_mul(BSIZE) {
             return -1;
         }
         let mut tot: u32 = 0;
@@ -503,7 +503,7 @@ pub const FSMAGIC: i32 = 0x10203040;
 pub const NDIRECT: i32 = 12;
 
 pub const NINDIRECT: i32 = BSIZE.wrapping_div(mem::size_of::<u32>()) as i32;
-pub const MAXFILE: i32 = NDIRECT.wrapping_add(NINDIRECT);
+pub const MAXFILE: usize = (NDIRECT as usize).wrapping_add(NINDIRECT as usize);
 
 /// Inodes per block.
 pub const IPB: i32 = BSIZE.wrapping_div(mem::size_of::<Dinode>()) as i32;
